Polishing up packaging

+++ Rakefile
@@ -33,15 +33,33 @@ Rake::VersionTask.new
 
 
 
+# Packaging
+#
+# Based on Travelling Ruby and FPM:
+# - http://phusion.github.io/traveling-ruby
+# - https://github.com/jordansissel/fpm
+#
+# I was able to archive my own extensions by packaging the software on the
+# target platform with the target-specific options (NO_EXT=true DIR_ONLY=true)
+# then extracting the resulting gem and its extensions into a tarball:
+#
+# target="linux-x86_64" # or "osx"
+# for ext in snappy-0.0.11 eventmachine-1.0.5 ; do
+#   name="${ext%%\-*}"
+#   version="${ext##*\-}"
+#   save="$(pwd)/$name"
+#   out="$(pwd)/$ext-$target.tar.gz"
+#   pushd franz/vendor/ruby
+#     find */extensions/*/*/$name* > $save
+#     find */gems/$name* >> $save
+#     tar -cvzf $out `cat $save`
+#   popd
+#   rm -f $save
+# done
+#
+require_relative 'lib/franz/metadata'
 
-require 'franz'
-
-TRAVELING_RUBY_VERSION = '20150210-2.1.5'
-SNAPPY_VERSION = '0.0.11'
-EM_VERSION = '1.0.5'
-
-desc 'Package Franz into binaries'
-task package: %w[ package:linux package:osx ]
+include Franz
 
 namespace :package do
   desc 'Package Franz for Linux (x86_64)'
@@ -77,11 +95,13 @@ namespace :package do
       sh "mv pkg/tmp/vendor pkg"
     end
     sh "rm -rf pkg/tmp"
-    sh "rm -f pkg/vendor/*/*/cache/*"
-    sh "rm -rf pkg/vendor/ruby/*/extensions"
-    sh "find pkg/vendor/ruby/*/gems -name '*.so' | xargs rm -f"
-    sh "find pkg/vendor/ruby/*/gems -name '*.bundle' | xargs rm -f"
-    sh "find pkg/vendor/ruby/*/gems -name '*.o' | xargs rm -f"
+    if !ENV['NO_EXT']
+      sh "rm -f pkg/vendor/*/*/cache/*"
+      sh "rm -rf pkg/vendor/ruby/*/extensions"
+      sh "find pkg/vendor/ruby/*/gems -name '*.so' | xargs rm -f"
+      sh "find pkg/vendor/ruby/*/gems -name '*.bundle' | xargs rm -f"
+      sh "find pkg/vendor/ruby/*/gems -name '*.o' | xargs rm -f"
+    end
   end
 end
 
@@ -114,35 +134,42 @@ file "pkg/eventmachine-#{EM_VERSION}-osx.tar.gz" do
 end
 
 def create_package target
-  package_dir = "franz-#{Franz::VERSION}-#{target}"
-  sh "rm -rf #{package_dir}"
+  package_name = "franz-#{VERSION}-#{target}"
+  package_file = ::File.join Dir.pwd, 'pkg', "#{package_name}.tar.gz"
+  package_dir = ::File.join Dir.pwd, 'pkg', package_name
+  output = ::File.join Dir.pwd, 'pkg', "franz_#{VERSION}_amd64.deb"
+  sh "rm -rf #{package_dir} #{output}"
   sh "mkdir -p #{package_dir}/franz"
   sh "cp -R bin #{package_dir}/franz"
   sh "mkdir #{package_dir}/franz/ruby"
   sh "tar -xzf pkg/traveling-ruby-#{TRAVELING_RUBY_VERSION}-#{target}.tar.gz -C #{package_dir}/franz/ruby"
-  sh "cp pkg/franz.sh #{package_dir}/franz.sh"
+  sh "cp pkg/franz.sh #{package_dir}"
   sh "cp -pR pkg/vendor #{package_dir}/franz/vendor"
   sh "cp -R franz.gemspec Readme.md LICENSE VERSION Gemfile Gemfile.lock lib #{package_dir}/franz/vendor"
   sh "mkdir #{package_dir}/franz/vendor/.bundle"
   sh "cp pkg/bundler-config #{package_dir}/franz/vendor/.bundle/config"
-  sh "tar -xzf pkg/snappy-#{SNAPPY_VERSION}-#{target}.tar.gz -C #{package_dir}/franz/vendor --strip 1"
-  sh "tar -xzf pkg/eventmachine-#{EM_VERSION}-#{target}.tar.gz -C #{package_dir}/franz/vendor --strip 1"
-  sh %Q~
-    which lsb_release || exit 0
-    which fpm || exit 0
-    fpm --verbose \
-      -s dir -t deb -C #{package_dir} \
-      -n franz -v #{Franz::VERSION} \
-      --license ISC \
-      --description "Franz" \
-      --maintainer "Sean Clemmer <sclemmer@bluejeans.com>" \
-      --vendor "Blue Jeans Network" \
-      franz.sh=/usr/local/bin/franz \
-      franz=/opt
-  ~
+  if !ENV['NO_EXT']
+    sh "tar -xzf pkg/snappy-#{SNAPPY_VERSION}-#{target}.tar.gz -C #{package_dir}/franz/vendor/ruby"
+    sh "tar -xzf pkg/eventmachine-#{EM_VERSION}-#{target}.tar.gz -C #{package_dir}/franz/vendor/ruby"
+  end
+  if !ENV['NO_FPM'] && target =~ /linux/
+    sh %Q~
+      fpm --verbose \
+        -s dir -t deb -C #{package_dir} \
+        -n franz -v #{VERSION} \
+        --license "#{LICENSE}" \
+        --description "#{SUMMARY}" \
+        --maintainer "#{AUTHOR} <#{EMAIL}>" \
+        --vendor "#{AUTHOR}" \
+        --url "#{HOMEPAGE}" \
+        --package "#{output}" \
+        franz.sh=/usr/local/bin/franz \
+        franz=/opt
+    ~
+  end
   if !ENV['DIR_ONLY']
-    sh "tar -czf #{package_dir}.tar.gz #{package_dir}"
-    sh "rm -rf #{package_dir}"
+    sh "cd #{package_dir} && tar -czf #{package_file} ."
+    sh "rm -rf #{package_dir} pkg/vendor"
   end
 end
